Feat:게시물 삭제 여부

+++ src/pages/Post/PostDetail.jsx
@@ -91,8 +91,10 @@ const PostDetail = () => {
     navigate(`/post/update/${postId}`);
   };
 
-  const deletePost = () => {
-    dispatch(__deletePost(postId));
+  const deletePost = async () => {
+    if (window.confirm("게시글을 삭제합니다")) {
+      dispatch(__deletePost(postId));
+    } else return;
   };
 
   const ZZim = async (e) => {
